+++ tests/tokens.rs
@@ -1,4 +1,4 @@
-use rhai::{Engine, ParseErrorType};
+use rhai::{Engine, EvalAltResult, ParseErrorType, RegisterFn, INT};
 
 #[test]
 fn test_tokens_disabled() {
@@ -17,4 +17,33 @@ fn test_tokens_disabled() {
         *engine.compile("let x = 40 + 2; x += 1;").expect_err("should error").0,
         ParseErrorType::BadInput(ref s) if s == "Unexpected '+='"
     ));
+}
+
+#[test]
+fn test_tokens_custom_operator() -> Result<(), Box<EvalAltResult>> {
+    let mut engine = Engine::new();
+
+    // Register a custom operator called `foo` and give it
+    // a precedence of 140 (i.e. between +|- and *|/).
+    engine.register_custom_operator("foo", 140).unwrap();
+
+    // Register a binary function named `foo`
+    engine.register_fn("foo", |x: INT, y: INT| (x * y) - (x + y));
+
+    assert_eq!(
+        engine.eval_expression::<INT>("1 + 2 * 3 foo 4 - 5 / 6")?,
+        15
+    );
+
+    assert_eq!(
+        engine.eval::<INT>(
+            r"
+                fn foo(x, y) { y - x }
+                1 + 2 * 3 foo 4 - 5 / 6
+            "
+        )?,
+        -1
+    );
+
+    Ok(())
 }