+++ RELEASES.md
@@ -6,8 +6,9 @@ Version 0.17.0
 
 This version adds:
 
-* [`serde`](https://crates.io/crates/serde) support for working with `Dynamic` values (particularly _object maps_)
-* Ability to surgically disable keywords and/or operators in the language
+* [`serde`](https://crates.io/crates/serde) support for working with `Dynamic` values (particularly _object maps_).
+* Ability to surgically disable keywords and/or operators in the language.
+* Ability to define custom operators (which must be valid identifiers).
 
 Breaking changes
 ----------------
@@ -20,6 +21,7 @@ New features
 * New `serde` feature to allow serializating/deserializating to/from `Dynamic` values using [`serde`](https://crates.io/crates/serde).
   This is particularly useful when converting a Rust `struct` to a `Dynamic` _object map_ and back.
 * `Engine::disable_symbol` to surgically disable keywords and/or operators.
+* `Engine::register_custom_operator` to define a custom operator.
 
 
 Version 0.16.1